removing api functions from codeforces util
trying to center all acesses to codeforces api
